requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_marshmallow
flask_cors
marshmallow-sqlalchemy
psycopg2
python-dotenv